Add more fields needed by entur-clients, make API more consistent
