--- flow-types/Authority.js
@@ -1,9 +1,0 @@
-// @flow
-
-export type Authority = {
-    id: string,
-    name: string,
-    codeSpace: string,
-    url?: string,
-    description?: string,
-}
