--- flow-types/Operator.js
@@ -1,7 +1,0 @@
-// @flow
-
-export type Operator = {
-    id: string,
-    name: string,
-    url?: string,
-}
