--- flow-types/Quay.js
@@ -1,7 +1,0 @@
-// @flow
-export type Quay = {
-  id: string,
-  name: string,
-  description?: string,
-  publicCode?: string,
-}
