--- flow-types/Situation.js
@@ -1,25 +1,0 @@
-// @flow
-
-import type { MultilingualString } from './MultilingualString'
-
-type ValidityPeriod = {
-    startTime: string,
-    endTime: string,
-}
-
-export type ReportType = 'general' | 'incident' | null
-
-type InfoLink = {
-    uri: string,
-    label: string,
-}
-
-export type Situation = {|
-    situationNumber: string,
-    summary: Array<MultilingualString>,
-    description: Array<MultilingualString>,
-    detail: Array<MultilingualString>,
-    validityPeriod: ValidityPeriod,
-    reportType: ReportType,
-    infoLinks: Array<InfoLink>,
-|}
