+++ flow-types/index.js
@@ -1,21 +1,11 @@
 // @flow
-export type { Authority } from './Authority'
 export type { BikeRentalStation } from './BikeRentalStation'
 export type { Coordinates } from './Coordinates'
 export type { DestinationDisplay } from './DestinationDisplay'
 export type { Feature, Category } from './Feature'
-export type { EstimatedCall, IntermediateEstimatedCall } from './EstimatedCall'
-export type { Leg } from './Leg'
-export type { Line } from './Line'
 export type { Location } from './Location'
 export type { TransportMode, LegMode, TransportSubmode } from './Mode'
 export type { MultilingualString } from './MultilingualString'
 export type { TypeName, NearestPlace } from './NearestPlace'
-export type { Notice } from './Notice'
 export type { PointsOnLink } from './PointsOnLink'
-export type { Place } from './Place'
-export type { Quay } from './Quay'
-export type { JourneyPattern, ServiceJourney } from './ServiceJourney'
-export type { Situation, ReportType } from './Situation'
 export type { StopPlace } from './StopPlace'
-export type { TripPattern } from './TripPattern'
