+++ index.d.ts
@@ -165,8 +165,9 @@ export interface Operator {
 export interface Quay {
     id: string;
     name: string;
-    description?: string;
-    publicCode?: string;
+    description: string;
+    publicCode: string;
+    situations: Array<Situation>;
 }
 
 export interface StopPlace {
@@ -221,16 +222,6 @@ export interface DestinationDisplay {
 }
 
 export interface EstimatedCall {
-    date: string;
-    destinationDisplay: DestinationDisplay;
-    forAlighting: boolean;
-    forBoarding: boolean;
-    situations: Array<Situation>;
-    notices?: Array<Notice>;
-    requestStop: boolean;
-}
-
-export interface IntermediateEstimatedCall {
     actualArrivalTime?: string; // Only available AFTER arrival has taken place
     actualDepartureTime?: string; // Only available AFTER departure has taken place
     aimedArrivalTime: string;
@@ -242,11 +233,15 @@ export interface IntermediateEstimatedCall {
     expectedDepartureTime?: string; // Only available BEFORE departure has taken place
     forAlighting: boolean;
     forBoarding: boolean;
-    situations: Array<Situation>;
     notices?: Array<Notice>;
     quay?: Quay;
+    realtime: boolean;
     requestStop: boolean;
+    serviceJourney: ServiceJourney;
+    situations: Array<Situation>;
 }
+
+export type IntermediateEstimatedCall = EstimatedCall
 
 export type BookingMethod = 'callOffice' | 'online';
 
@@ -267,12 +262,15 @@ export interface BookingArrangement {
 }
 
 export interface Line {
+    bookingArrangements?: BookingArrangement;
+    description?: string;
+    flexibleLineType?: FlexibleLineType;
     id: string;
     name: string;
     notices?: Array<Notice>;
     publicCode: string;
-    bookingArrangements?: BookingArrangement;
-    flexibleLineType?: FlexibleLineType;
+    transportMode: TransportMode;
+    transportSubmode: TransportSubmode;
 }
 
 export type FlexibleLineType =
@@ -404,36 +402,9 @@ export interface NearestPlace {
  * Stop Place
  */
 
-export interface Departure {
-    date: string;
-    destinationQuay?: Quay;
-    forBoarding: boolean;
-    requestStop: boolean;
-    forAlighting: boolean;
-    destinationDisplay: DestinationDisplay;
-    notices?: Array<Notice>;
-    aimedDepartureTime: string;
-    expectedDepartureTime: string;
-    realtime: boolean;
-    situations?: Array<Situation>;
-    cancellation: boolean;
-    quay: Quay;
-    serviceJourney: ServiceJourney & {
-        line: Line & {
-            transportMode: TransportMode;
-            description?: string;
-        };
-    };
-}
-
-export interface QuayDepartures {
+export interface DeparturesById {
     id: string;
-    departures: Array<Departure>;
-}
-
-export interface StopPlaceDepartures {
-    id: string;
-    departures: Array<Departure>;
+    departures: Array<EstimatedCall>;
 }
 
 export interface StopPlaceDetails {
@@ -582,23 +553,23 @@ declare class EnturService {
   getDeparturesFromStopPlaces(
       stopPlaceIds: Array<string>,
       params?: GetDeparturesParams,
-  ): Promise<Array<StopPlaceDepartures | undefined>>;
+  ): Promise<Array<DeparturesById | void>>;
 
   getDeparturesFromStopPlace(
       stopPlaceId: string,
       params?: GetDeparturesParams,
-  ): Promise<Departure[]>;
+  ): Promise<DeparturesById[]>;
 
   getDeparturesFromQuays(
       quayIds: Array<string>,
       params?: GetDeparturesParams,
-  ): Promise<Array<QuayDepartures | undefined>>;
+  ): Promise<Array<DeparturesById | undefined>>;
 
   getDeparturesBetweenStopPlaces(
       fromStopPlaceId: string,
       toStopPlaceId: string,
       params?: GetDeparturesBetweenStopPlacesParams,
-  ): Promise<Departure[]>;
+  ): Promise<EstimatedCall[]>;
 
   getNearestPlaces(
       coordinates: Coordinates,
