+++ src/fields/Interchange.js
@@ -1,0 +1,11 @@
+// @flow
+
+export type Interchange = {
+    guaranteed: boolean,
+    staySeated: boolean,
+}
+
+export default {
+    guaranteed: true,
+    staySeated: true,
+}
