+++ src/fields/Notice.js
@@ -3,3 +3,7 @@
 export type Notice = {|
     text: string,
 |}
+
+export default {
+    text: true,
+}
