+++ src/fields/PointsOnLink.js
@@ -1,0 +1,11 @@
+// @flow
+
+export type PointsOnLink = {|
+    points: string,
+    length: number,
+|}
+
+export default {
+    points: true,
+    length: true,
+}
