+++ src/fields/Quay.js
@@ -1,0 +1,19 @@
+// @flow
+
+import situationFields, { type Situation } from './Situation'
+
+export type Quay = {|
+  id: string,
+  name: string,
+  description: string,
+  publicCode: string,
+  situations: Array<Situation>,
+|}
+
+export default {
+    id: true,
+    name: true,
+    publicCode: true,
+    description: true,
+    situations: situationFields,
+}
